fixup: Moved Store::ParsingBuffer to dedicated files
