+++ src/store/Makefile.am
@@ -22,6 +22,8 @@ libstore_la_SOURCES = \
 	Disks.h \
 	LocalSearch.cc \
 	LocalSearch.h \
+	ParsingBuffer.cc \
+	ParsingBuffer.h \
 	Storage.h \
 	SwapMeta.cc \
 	SwapMeta.h \
